app/requirements.txt
flask>=3.0,<4
azure-containerregistry>=1.2,<2
azure-identity>=1.16,<2
